move configuration to AppLoader and move secret to dynamo config
