+++ app/controllers/Application.scala
@@ -1,15 +1,16 @@
 package controllers
 
 import com.typesafe.config.Config
+import play.api.Configuration
 import play.api.mvc._
 
 //noinspection TypeAnnotation
-class Application (cc: ControllerComponents, underlyingConfig: Config, documentation: () => Seq[(String, String, String)]) extends AbstractController(cc) {
+class Application (cc: ControllerComponents, configuration: Configuration, documentation: () => Seq[(String, String, String)]) extends AbstractController(cc) {
   def index = Action {
     Ok(views.html.index(documentation()))
   }
 
   def config = Action {
-    Ok(underlyingConfig.root().render())
+    Ok(configuration.underlying.root().render())
   }
 }
