+++ conf/application.conf
@@ -5,7 +5,8 @@
 # ~~~~~
 # The secret key is used to secure cryptographics functions.
 # If you deploy your application to several instances be sure to use the same key!
-play.http.secret.key="jA7Zs?VuTmnL[`JeMm05<:lK2sx3odM>Dds]43OSb2?RfRAuI2/X=?DhBi/PBovg"
+# This is set in the Dynamo configuration
+# play.http.secret.key=""
 
 # The application languages
 # ~~~~~
